requirements.txt
django==4.2.7
django-async-redis==0.2.0
django-cors-headers==4.3.1
django-filter==23.5
django-haystack==3.2.1
django-haystack-es==0.2.8
django-money==3.4.1
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
celery==5.3.6
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
haystack==0.42
Flask==2.0.1
requests==2.26.0
pytest==6.2.4
daphne==4.0.0
channels==4.0.0
django-money==3.2.0
fcm-django==0.3.8
drf-extensions==0.7.1
drf-spectacular==0.26.3
drf-spectacular-sidecar==0.0.5
psycopg2-binary==2.9.7
redis==5.0.4
autobahn==23.6.2
firebase-admin==6.5.